[network] optimize ethtool waiting time for down interfaces
Waiting for interfaces to be up can be long in some cases (like if STP
is used on the switch). Since we cannot known in advance if an interface
will eventually be up or down (e.g: no carrier) a retry loop with a
timeout was used. This was causing delay for clusters where interfaces
were coming up fast since we had to wait the end of the timeout to act
that down interfaces were "really down", without being to optimistic and
miss interfaces that were supposed to be up.

By using the reference api file presents on each node, we can predict
that an interface is supposed to have no carrier and skip the waiting
loop. This is not applicated to -m api mode which needs to be certain of
interfaces status.