[ohai/g5k] add an entry containing the local api description
This allow usage of information contained inside the reference-api for a
node, using the file /etc/grid5000/ref-api.json which is populated by
g5k-postinstall.
It can be usefull inside ohai plugins to have reference api data to make
conditionnals treatment.
If g5k-checks is running in api mode or if the file does not exist, the
entry will have the value nil.